+++ applications/platform/designer/src/ovdCPlayerVisualization.cpp
@@ -682,7 +682,7 @@ void CPlayerVisualization::drag_data_get_from_widget_cb(GtkWidget* pSrcWidget, G
 {
 	char l_sString[1024];
 	sprintf(l_sString, "%p", pSrcWidget);
-	gtk_selection_data_set_text(pSelectionData, l_sString, strlen(l_sString));
+	gtk_selection_data_set_text(pSelectionData, l_sString, gint(strlen(l_sString)));
 }
 
 void CPlayerVisualization::drag_data_received_in_widget_cb(GtkWidget* pDstWidget, GdkDragContext* /*pDragContext*/, gint /*iX*/, gint /*iY*/, GtkSelectionData* pSelectionData, guint /*uiInfo*/, guint /*uiTime*/, gpointer /*pData*/)
