+++ libraries/lib-advanced-visualization/src/mCRendererMountain.cpp
@@ -56,7 +56,7 @@ void CRendererMountain::rebuild(const IRendererContext& rContext)
 		for (j = 0; j < m_sampleCount - 1; ++j)
 		{
 			const size_t id = k * 6;
-			const uint32_t v1 = i * m_sampleCount + j;
+			const uint32_t v1 = uint32_t(i * m_sampleCount + j);
 			const uint32_t v2 = v1 + m_sampleCount;
 			m_oMountain.m_vTriangle[id] = v1;
 			m_oMountain.m_vTriangle[id + 1] = v2;
@@ -82,7 +82,7 @@ void CRendererMountain::refresh(const IRendererContext& rContext)
 	for (i = 0, k = 0; i < rContext.getSelectedCount(); ++i)
 	{
 		k = ((m_historyCount - 1) / m_sampleCount) * m_sampleCount;
-		std::vector<float>& l_vHistory = m_history[rContext.getSelected(i)];
+		std::vector<float>& l_vHistory = m_history[rContext.getSelected(uint32_t(i))];
 		CVertex* l_pVertex = &m_oMountain.m_vVertex[i * m_sampleCount];
 		for (size_t j = 0; j < m_sampleCount; j++, k++)
 		{
